Global TRISA Directory Service Implementation
This commit moves the v1beta1 directory service implementation
from the trisacrypto/testnet repository to begin creating official
releases of the directory service. The GDS implementation is not
a concern of a TRISA client or peer and therefore is stored in its
own repository to prevent confusion. Extra services such as the
admin or replication service are added here for convienience.