Backup Manager (#5)
Implements a Backup manager routine that copies the leveldb into a tar.gz archive on disk every 24 hours (or at another specified interval). The Backup manager can be disabled, but if enabled, only keeps a specific number of backups (defaults to 1). At some point, we should probably move the backup file to cloud storage, but this is a good start to unblock other directory service updates.

Additionally, this PR updates the trisads package name to gds and enhances the gds configuration module.