Registration Form (#7)
Re-implements the registration form with basic functionality and the ability to download the JSON data required to make a registration. Intended as a temporary hold over for copy review until we can hook up the backend.

Also implements a docker-compose functionality in order to develop against the front-end and alerts handling for errors when using the directory service.