GDS UI Local Storage and Load/Download
Enhances the GDS UI registration form with local storage so that the
form persists between browser tabs and when the browser is reloaded,
making the long form easier to fill out. It also adds a reset button the
summary page to clear the form out and start over.

Additionally adds a download button to download the JSON data for the
form and a load button to load downloaded data back into the form. This
makes the form usable for editing registration forms before the database
is live.