Implement changes from trisa v1beta1 pb update
This is a lot of changes that gets the code to compile, but it has not
been checked for correctness. The tests pass, but there are limited
tests in this package. This code cannot be deployed until the database
is migrated. We should either continue working off this branch or merge
a broken version, tagging it as such.