Refactor emails to email manager
Email handling was a mess and it was difficult to modify emails so I
refactored the email manager to its own package. This simplified things
a lot and I created tests to make sure that I could review only the
email creation and sending process. A review of the interaction with the
server code is still required.