GDS Web UI v1beta1 Release (#18)
* update docker-compose to run latest GDS server

* tab links and urls on main page

* route navigation with 404 and verify contact stub

* generate v1beta1 js protobufs

* verify contact web ui

* cleanup of gdsutil (formerly debug) for testing UI

* better lookup results

* country.js attribution

* Refactor emails to email manager

Email handling was a mess and it was difficult to modify emails so I
refactored the email manager to its own package. This simplified things
a lot and I created tests to make sure that I could review only the
email creation and sending process. A review of the interaction with the
server code is still required.

* fix email whitespace

* better error messaging returned from server

* fix verified on bug

* submit registration form