Better contact handling in email client senders (#36)
Performs better handling of sending emails to technical, admin, billing,
and legal contacts. First, a check is added to ensure the contact is not
nil and that it has an email address, otherwise it is skipped. Sending
does not stop on the first error, but all email addresses are tried to
make sure at least one email can get out. The number of errors are
reported via logging rather than returned. If no messages were delivered
then the function returns an error.

This PR also updates the Registration Review email to be nicely rendered
YAML with string enums and no empty fields.

Fixes #31
Fixes #32