Reindex and gzip index sync without DB in maintenance mode (#52)
* Reindex and index sync

Updates the database configuration to accept a bool for reindexing on
boot and the store reindexes if necessary on open. Adds a call to sync
to on Backup so that the indices are resync'd routinely.

Fixes #30

* No DB on maintenance mode, compressed ldb indices

This commit enables more compact LevelDB indices by compressing them
before writing them to disk. However, this requires deleting old indices
and regenerating them. To facilitate this, during maintenance mode, the
database no longer opens a connection to the database or runs any
management processes. This will allow us to ssh into the server and
access the database as needed to perform any changes.

Fixes #47

* check error on reindex