Implements Resend Admin RPC (#54)
* Implements Resend Admin RPC

There have been a couple issues with email deliveries over the first few
registrations, requiring that emails be resent. This RPC allows an admin
to trigger a resend of different email types, based on the condition of
the VASP record. This commit also tracks the number of emails that are
sent for logging purposes.