Gossip Cleanup
This PR finalizes the v1.1 Gossip implementation though it is not ready
for production release quite yet. Our intent is to run GDS with the
replica server disabled but to have object versioning in place. The
cleanup PR creates the ObjectStore abstraction to remove directly use of
leveldb and implements the routine that initiates Gossip with remote
replicas.