Gossip cleanup (#68)
* Gossip Cleanup

This PR finalizes the v1.1 Gossip implementation though it is not ready
for production release quite yet. Our intent is to run GDS with the
replica server disabled but to have object versioning in place. The
cleanup PR creates the ObjectStore abstraction to remove directly use of
leveldb and implements the routine that initiates Gossip with remote
replicas.

* AntiEntropy background routine

Implements the anti-entropy background routine in the Replica with a
jitter ticker that adds random noise to the anti-entropy interval.