Differentiate TestNet vs Production UI and Emails
This PR updates the GDS-UI with an environment variable that determines
at build time if the UI is TestNet or Production then sets constants to
change the text and some visual markers. The GDS-UI also gets a bit of a
facelift.

The emails sent by the GDS are also modified so that recipients can tell
which directory is sending them emails.