WIP: Admin Status RPC (#76)
The Admin server Status RPC returns the count of states for all VASP
registrations and Certificate Requests. This PR also implements
ListVASPs on the Store object in order to support the Status endpoint
and refactors the Store List* methods to use an iterator rather than
materialize the entire list in memory.