Shared Secret Configuration for Admin JWT Tokens
Implements rotating RSA secret keys for use in signing JWT tokens for
authentication with the admin API. Asymmetric signatures are used so
that different services can verify a token using the public keys.
Multiple tokens are implemented in the configuration to allow keys to be
rotated regularly. A gdsutil function was created to generate keys along
with with an associated ksuid as the key id.

Fixes SC-563