Refactor Admin gRPC API to JSON API (#90)
Refactors the gRPC Admin API into a JSON API that is better suited to the GDS Admin UI and other administrative tasks with Google Authentication. This PR implements the new Admin API client definitions and modifies the main.go file accordingly. It also sets up an http server to listen on the admin port and uses the Gin framework to implement a JSON API.