Enable local development with Docker Compose Part 2 (#107)
This PR is intended to facilitate local development with docker compose
and makes some fairly dramatic changes that shouldn't affect production.
The biggest change is the move of the secret package from the main GDS
package into a subpackage that includes a Mock so that Google
application credentials are not required. Additionally, if Sectigo
credentials are not supplied, the server issues an error but does not
stop starting up. Finally environment variable templates, test keys, and
fixtures directories are created.