Implements Double Cookie Middleware
Implements the double cookie middleware for CSRF protection in the Admin
API. The CSRF protection relies on a cookie that is HttpOnly, e.g. one
that cannot be read in JavaScript and one that can be. The client
responsibility is to read the cookie that can be read and submit it in
an X-CSRF-TOKEN header with the request.