Simple CLI Authentication (#129)
To ensure the CLI still works on the MVP release, this PR adds private
key based authentication to the CLI to avoid having to go through the
OAuth2 workflow. This mechanism will be replaced in the future but is a
stop gap for now. This commit also fixes a reauthentication bug.