Edit client profiles from the command line
This extension allows you to use --edit with the gds profiles, gdsutil
profiles, and trtl profiles command to directly edit the profiles YAML
in the command line with vim, emacs, or nano. You can specify the editor
of your choice using the $EDITOR environment variable. Note however that
a forking editor like VSCode will not work with this CLI app right now.