BugFix: Common Name Validation
It appears that the semantics of net.SplitHostPort have changed and now
if a port is not included in an addr, then an error is returned. This
prevents us from using this method for common name validation and the
errors have started showing up in the UI. The new method of validation
uses a regular expression, which isn't great, but it does do the job for
all of the cases provided in the tests.